+++ ChoGGi's Library/Changelog.md
@@ -3,6 +3,7 @@
 ## v6.4 (Unreleased)
 - ValidateImage didn't work with non-Library images.
 - DeleteObject works with multi-selection.
+- DotNameToObject didn't work with spaces.
 - Removed some log spam from:
 - > DeleteObject
 - > List Choice dialog.